requirements.txt
mysql.connector
pytest